Renaming test spec for end to end test contentful
I want to align spec names with github repo names.
