+++ spec/contentful/contentful_spec_helper.rb
